requirements.txt
beautifulsoup4==4.12.3
matplotlib==3.9.2
numpy==2.1.3
pandas==2.2.3
Requests==2.32.3
seaborn==0.13.2
st_annotated_text==4.0.1
streamlit==1.40.1
transformers==4.46.1
Unidecode==1.3.8
